lxc: Initializing IPv6 and IPv4 gateway to overwrite old settings.
This commit fixes a bug when you have multiple network settings defined.
Basically, if you set an IPv6 or IPv4 gateway, it carries on next
network settings. It is happening because the data is not being
initialized when a new network type is defined. So, the old data still
persists into the pointer.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <20181128033857.30303-2-jcfaracco@gmail.com>