qemu_process: All ProcessQMP errors are fatal
In the past capabilities could be determined in other ways if QMP
messaging didn't succeed so a non-fatal error case was included in the
capabilities and QMP Process code.

For a while now, QMP capabilities failure has been a fatal case.

This patch makes QMP process failures return as a fatal error in all
cases consistent with 1) all failures actually being fatal in
QMP capabilities code and 2) the QMP process code being made generic.

The process changes impact the capabilities code because non-fatal
return codes are no longer returned.

The rest of the QMP associated capabilities code is updated to make all
errors fatal for consistency.

The process changes also force a change in QMP process stderr reporting
because the previous triggers for stderr reporting are no longer passed
through the function interfaces.

As best I can tell the only case where QMP process stderr should be
explicitly reported is when the QMP process exits with a non-zero
status.

Error reporting is moved to virQEMUCapsInitQMP and the QMP process
stderr is reported only when the QMP process status code is non-zero
(and is only reported for the primary QMP query not the secondary TCG
 specific QMP query.)

Signed-off-by: Chris Venteicher <cventeic@redhat.com>
Message-Id: <20181203051030.22126-9-cventeic@redhat.com>