qemu_process: Persist stderr in qemuProcessQmp struct
A qemuProcessQmp struct tracks the entire lifespan of a single QEMU Process
including storing error output when the process terminates or activation
fails.

Error output remains available until qemuProcessQmpFree is called.

The qmperr variable is renamed stderr (captures stderr from process.)

The stderr buffer no longer needs to be maintained outside of the
qemuProcessQmp structure because the structure is used for a single QEMU
process and the structures can be maintained as long as required
to retrieve the process error info.

Signed-off-by: Chris Venteicher <cventeic@redhat.com>
Message-Id: <20181203051030.22126-10-cventeic@redhat.com>