qemuDomainRemoveRNGDevice: Remove associated chardev too
https://bugzilla.redhat.com/show_bug.cgi?id=1656014

An RNG device can consists of more devices than RND device
itself. For instance, in case of EGD there is a chardev that
connects to EGD daemon and feeds the qemu with random data. When
doing RNG device removal we have to remove the associated chardev
as well.

At the same time, we should not do any 'goto cleanup' until
virDomainAuditRNG() is called as it might result in us not
propagating the audit message.

Signed-off-by: Michal Privoznik <mprivozn@redhat.com>
Message-Id: <33f71e006ce5c614a19403a1cbf2903698b2a462.1543930755.git.mprivozn@redhat.com>