lxc: stop incorrectly validating interface type
Commit 017dfa27d changed a few switch statements in the LXC code to
have all possible enum values, and in the process changed the switch
statement in virLXCControllerGetNICIndexes() such that it returned
error status for any interfaces that weren't implemented with a veth
pair when it should have just been ignoring those interfaces.

Since the interface type will have already been validated before
reaching this function, we shouldn't be doing any validation at all -
just add the ifindex of the parent veth if its a veth pair, and ignore
it otherwise.

Resolves: https://bugzilla.redhat.com/1656463
Signed-off-by: Laine Stump <laine@laine.org>
Message-Id: <20181206023513.2912-2-laine@laine.org>