tests: Introduce negative versions of DO_TEST_CAPS_LATEST
As commit d8266ebe161 demonstrated, it's so easy to forget to add a
single capability which in turn can easily fool the test suite so that
tests expecting a failure can fail with a different error than we
expected, but still making those pass. Unfortunately for commit
d8266ebe161, it allowed a domain with multiple OpenGL-enabled graphics
devices to start successfully. As a precaution measure, introduce
negative versions of DO_TEST_CAPS_LATEST macros, so that we eliminate
this vector from now on.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <9f64207f7eaacaaf5c685246ba2e82c045f5a3db.1544193131.git.eskultet@redhat.com>