conf: domain: gfx: Iterate over graphics devices when doing validation
The validation code for graphics has been in place for a while, but
because it is only executed from the device iterator, that validation
code was never truly run. The unfortunate side effect of this whole mess
was that a few capabilities were missing from the test suite, which in
turn meant that a few graphics test which expected a failure happily
accepted whatever failure the parser returned which made them succeed
even though in reality we allowed to start a domain with multiple
OpenGL-enabled graphics devices.
This patch enables iteration over graphics devices. Unsurprisingly,
a few tests started failing as a result, so fix those too.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <0be995e7206828fb1083e24c332c998001d2813a.1544193131.git.eskultet@redhat.com>