qemu: snapshot: save/restore inactive persistent config
In case of active persistent domain snapshot metadata is
not complete. We save only active configuration and on
restore use it both for active and inactive configuration.
Let's fix it and save and restore both in this case.

In case of active transient or inactive domain we have
only one config and thus everything is good.

By the way this patch fixes config memleak on error paths.

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1544688239-814454-4-git-send-email-nshirokovskiy@virtuozzo.com>