qemu: Update a domain's persistent definition file if it's defined
During making disk snapshot in an active domain, sometimes we
should update the domain's persistent definition.We must check if
the domain is defined, before we update the persistent definition
file.
First,we create a vm.Then,define the vm and undefine the vm.Last create a
--memspec snapshot for the vm.Now,we will find a persistent definition
file in the config directory if we don't use this patch.

Signed-off-by: Li XueLei <li.xuelei1@zte.com.cn>
Message-Id: <1545296773-11773-1-git-send-email-wang.yechao255@zte.com.cn>