conf: snapshot: remove snapshot mode checking from disk align
The removed hunk of code added in [1] in virDomainSnapshotAlignDisks
either duplicates qemuDomainSnapshotPrepare checks or prohibit some
meaningful cases and overall is complicated to think about (check next analysis).

It prohibit next cases:

A. external disk snapshot, which fall in 3 cases

1. there are other disks with internal snapshot
2. all other disks snapshots are external and domain is active
3. the same as above but domain is inactive

Cases 1 and 2 are disabled by qemuDomainSnapshotPrepare too and case 3 need not
to be prohibited. This way one can make external disk snapshot of inactive
domain without disk only flag. One may argue why we need such a possibility
- then why we have code at the bottom of qemuDomainSnapshotPrepare that adds
  disk only to @flags. This @flags mutating code starts to function only in case 3.

B. disabling disk snapshot if it is not disabled in domain disk config too.

This actually does not make sense. Say snapshot is not disabled in disk config
and this check gives failure. Then I can disable it in disk config and pass
this check and after that disk config is never checked so it is not clear why
fail first case and pass second. I guess the intention was to prohibit
disabling non readonly disks as for readonly disk snapshot attribute is
set to none on parse. So we'd better analyze these possibilites which fall into
4 cases (all non-readonly):

1. empty disk and active domain
2. empty disk and inactive domain
3. non empty disk and there are some disks with internal snapshots
4. non empty disk and all other disks snapshots are external

Cases 1, 2 and 4 are handled valid cases. The reasoning for case 2 is
a bit tricky - it is hanled because only floppies are possible as
empty non readonly disks and those are skipped by qemuDomainSnapshotForEachQcow2Raw.
Case 3 is prohibited by qemuDomainSnapshotPrepare.

Now @require_match in virDomainSnapshotAlignDisks is not used and can be
removed.

The above analysis only covers actually creating snapshot but not redefining.
Reasoning here is just as in the previous patch - I guess users are not
supposed to edit disk's snapshot attributes.

[1] f9670b - snapshot: improve disk align checking

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1545906052-101494-6-git-send-email-nshirokovskiy@virtuozzo.com>