qemu: disable internal snapshot of readonly disk
Now in case of active domain such a snapshot will complete
successfully but disk will not have a correspondent internal snapshot
as qemu will not make one. In case of inactive domain operation
will finish successfully too but disk snapshot will be done.

Let's disable such a snapshot because as described above in case
of active domain it is not possible now and in both cases it does
not make much sense. We recently disable external snapshot of
readonly disk too in [1].

[1] 067aad26b - qemu: disable external snapshot of readonly disk

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1545906052-101494-8-git-send-email-nshirokovskiy@virtuozzo.com>