qemu: snapshot: fix for case of disk with empty source
It is possible to specify any snapshot option for disk with empty
source. Only cdrom and floppy can have empty source but cdrom
is out of considering because it is always readonly (set on parse
stage) and snapshotting readonly disks is now completely prohibited.
Let's check floppies.

In case of internal snapshot of such a disk for active or inactive domain
snapshot returns success but actual snapshot is not possible of course.
We need to prohibit this cases explicitly.

In case of external snapshot of such a disk for active or inactive domain
we get errors [1] and [2] respectively. It is better to have more
user-frieldly messages.

This patch prohibits explicitly all the above cases.

[1] error in case of external disk snapshot of active domain
error: internal error: unable to execute QEMU command 'transaction': Device 'drive-fdc0-0-0' has no medium

[2] error in case of external disk snapshot of inactive domain
error: internal error: Child process (/usr/local/bin/qemu-img create -f qcow2 \
       -o 'backing_file=(null),backing_fmt=qcow2' /path/file.qcow2) \
       unexpected exit status 1: qemu-img: /path/file.qcow2: \
       Could not open '/path/(null)': No such file or directory

Signed-off-by: Nikolay Shirokovskiy <nshirokovskiy@virtuozzo.com>
Message-Id: <1545906052-101494-9-git-send-email-nshirokovskiy@virtuozzo.com>