lxc: refactoring LXC network definition with a sparse array.
LXC was using a single data structure to define all LXC NICs. In terms
of optimization it is very interesting, but it is not useful when you
use a index to define networks. After major release 3.0, LXC adopted
indexes for network definitions. So, this commit adds a sparse vector to
handle network indexes. Now, networks can be defined in sequence using
the legacy config. Or using a sparse array, using version 3 with indexes.

Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <20181228190138.21128-2-jcfaracco@gmail.com>