vircgroup: introduce virCgroupV2DeviceAttachProg
This function loads the BPF prog with prepared map into kernel and
attaches it into guest cgroup.  It can be also used to replace existing
program in the cgroup if we need to resize BPF map to store more rules
for devices. The old program will be closed and removed from kernel.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <4b61da0f9304d94bb97e8c0ef346caabfc6efe20.1546437956.git.phrdina@redhat.com>