vircgroupv2: use dummy process to workaround kernel bug with systemd
If some program is attached to cgroup and that cgroup is removed before
detaching the program it might not be freed and will remain in the
system until it's rebooted.

This would not be that big deal to workaround if we wouldn't use
machined to track our guests.  Systemd tries to be the nice guy and if
there is no process attached to TransientUnit it will destroy the unit
and remove the cgroup from system which will hit the kernel bug.

In order to prevent this behavior of systemd we can move another process
into the guest cgroup and the cgroup will remain active even if we kill
qemu process while destroying guest.  We can then detach our BPF program
from that cgroup and call machined terminate API to cleanup the cgroup.

Signed-off-by: Pavel Hrdina <phrdina@redhat.com>
Message-Id: <b4028d19545e53afbb9b3e5e2387192cb624c581.1546437956.git.phrdina@redhat.com>