examples: Drop event-test.c dependency on gnulib <verify.h>
Pulling in gnulib just for the <verify.h> header is rather
expensive, especially since that header does not require us
to link against gnulib.  It's better to make the event-test
example be standalone by just open-coding a more limited form
of a verify() macro that depends on modern gcc (we have enough
CI coverage that even though the verify is now a no-op in
older setups, we will still notice if we fail to add an event
- as a quick test, I still provoked a compile failure on
Fedora 29 by deleting a line from domainEvents).

Signed-off-by: Eric Blake <eblake@redhat.com>
Message-Id: <20190108055842.19272-2-eblake@redhat.com>