conf: Introduce virStoragePoolXMLNamespace
Introduce the infrastructure necessary to manage a Storage Pool XML
Namespace. The general concept is similar to virDomainXMLNamespace,
except that for Storage Pools the storage backend specific details
can be stored within the _virStoragePoolOptions unlike the domain
processing code which manages its xmlopt's via the virDomainXMLOption
which is allocated/passed around for each domain.

This patch defines the add the parse, format, free, and href methods
required to process the XML and callout from the Storage Pool Source
parse, format, and clear/free API's to perform the action on the
XML data for/from the backend.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <20190108175226.32693-2-jferlan@redhat.com>